Login: prevent inactive accounts from triggering confirmation emails
Inactive AND unverified user accounts would trigger sending a verification
email upon login, which is utterly confusing for users. In this scenario
they would have the false perception that verifying their emails will
suffice to be able to log in, when the initial (and blocking) fact is an
admin deactivated their accounts.

They should instead be prevented from receiving any verification emails.