Update forms.py
Modified SignupForm and moved declaration of passwords inside __init__ method. The call to super() in __init__ places the password fields before any fields added in the __init__ of BaseSignupForm (e.g. email2) which is an undesired behaviour. The alternative is to have email2 in BaseSignupForm declared in __new__ (which might have other side effects).