feat(accounts): add key to password reset template
Currently, the password reset email template gets passed the full URL to
the password reset endpoint. However, in some scenarios, this endpoint
isn't the location where the user will actually perform the password
reset data entry, for example if Django only hosts an API backend and
the user-facing part instead is implemented via a SPA.

To make it easier to integrate such setups with allauth, this change
additionally passes the uid and key to the password reset email
template. This provides full control to the template author to generate
their own password reset links and support a split frontend/backend.